ai_core/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai==0.8.3